requirements.txt
--find-links https://flashinfer.ai/whl/cu124/torch2.5/flashinfer-python
pandas
colorama
openai
google-auth
google-cloud-aiplatform
requests
pyyaml
aisuite
gradio==5.5.0
sglang[all]>=0.4.3.post2
azure-ai-inference